Load and persist Vault attribute at initialization and save
Prior to this commit, each interaction with a vault_attribute would issue a
_new_ HTTP request to Vault and cache the result. This caused race conditions
and unexpected behaviors. Additionally, it was a very poor optimization that
turned out to actually be less performant that eager loading all the values
from Vault at once.

This commit changes vault-rails to decrypt all values from Vault after the
object is loaded, and then encrypt the values using Vault only right before
the object is to be saved.