:zap: Allow warning messages to be configurable
When a Rails application that uses Vault::Rails runs its test suite
using an in-memory cipher (a common use case), many hundreds of log
messages warning about using an insecure cipher in a production-like
environment can result. This increases the logging noise and reduces the
value of logging during testing.

Adding configuration methods to disable these in-memory warnings
improves the signal to noise ratio when using this gem in a test
environment.