Add support for Ruby 2.4.2
In Ruby 2.4.2, the OpenSSL library enforces key lengths for the
supported ciphers. In particular, the AES 128 cipher requires a key
length of 16 bytes. The previous behavior of the library was to silently
ignore invalid key lengths, but now an exception is raised. This change
limits the key length to 16 bytes.

While getting the Travis CI build to pass, I also dropped support for
Ruby 2.1 (EOL April 2017) and Rails 4.0 (EOL January 2015). I also
restricted the support of this gem to < 5.1 since 5.1 has breaking
changes to the ActiveRecord::Dirty module which this gem relies on.