Don't rely on Rails for the default encoding
If we've included vault-rails in an application that doesn't use rails but
only uses activerecord then we won't have access to
`Rails.application.config` to ask for `encoding`.  We wrap access to this
in some guard clauses.

Ideally we'd use the encoding on the database connection for the model, but
the `Vault::Rails.encrypt` and `Vault::Rails.decrypt` methods don't know
about models.  We could change these methods to take an optional encoding
parameter that the `Vault::Rails::EncryptedModel` will pass in from the
database connection.  However, we'd still need to work out a default for
the cases where we use these methods outside the context of a model, so
being more robust in the absence of a full rails application is good enough
for now.