Use `ActiveRecord::Base.logger` instead of `Rails.logger`
We only rely on activerecord now, not the full rails gem so it makes sense
to use the `ActiveRecord::Base.logger` instead of attempting to use the
`Rails.logger` if rails is available.  We don't need to guard against
`ActiveRecord::Base` being defined though, as if it's not we'll have much
bigger problems than whether or not we can use the logger.