Allow explicitly setting a default encoding value
If we can't rely on there being a full Rails app to get a default encoding
from, and we can't rely on the database connection for the models either
then it makes sense to allow setting an encoding explicitly for
Vault::Rails.  In most cases this will be excessive as we will be using it
in a rails app and can rely on that value, but for the few cases where we
can't this makes sense.  If the value is explicitly set we use that, if not
we fall back to the `Rails.application.config.encoding` value and if that
is not set (or not present because we're not in a rails app), we fall back
to a hardcoded default of UTF-8.

We've also made setting this value more robust by making sure that whatever
value is set is really a valid encoding.