Test against rails 5, 5.1, and 5.2 not just 4.2
We add the new rails gemfiles using the `appraisals` gem.  This means that
we can run `bundle exec appraisal rake` locally to run the specs against
rails 4.2, 5, 5.1, and 5.2.  We've also added these gemfiles to the
`.travis.yml` so these rails versions are added to the ruby version test
matrix that we run on CI.  Note that although we use CircleCI to run our
CI tests, we use the `wwtd` gem to read the `.travis.yml` config and do
what it would do.  This is confusing, but it works.