Fix bug with `vault_persist_before_save!` and rails 5.2
In rails 5.2 the behaviour of dirty tracking methods changed when they are
run during callbacks.  Methods like `<attribute>_changed?` and `changes`
do the same thing in callbacks run before save, but will behave differently
in callbacks run after save.  We'd already written
`__vault_encrypt_attribute` to understand this and use the
`saved_change_to_attribute?` method instead of `<attribute>_changed?`
method for rails 5.2.  Unfortunately, this is only the correct approach
when we're running `__vault_encrypt_attribute` in an `after_save` callback,
which we're not if we have run `vault_persist_before_save!`.  This means
`__vault_encrypte_attribute` is run in a `before_save` callback where the
`saved_change_to_attribute?` methods don't do what we want (because we
haven't saved any changes yet).  In this case we do want to use the old API
that the raisl < 5.2 branch uses.