Clear up Person vs. EagerPerson when testing vault_persist_before_save!
We use `Person` in the "when not used" context because that model class
has not had `vault_persist_before_save!` called on it.  `EagerPerson` has
and that's why we use it in the "when used" context.  It'd be clearer if
we could explicitly call it in the spec, but there's no clean up option to
unset it so we need separate models.