Extract attribute batch operations in PerformInBatches
For both encryption and decryption we need the same initialization phase of reading the
configuration options. The idea is to avoid code repetition and cluttering even more code in EncryptedModel.

Note that although this commit is forwardported from the 0.6 version we've
updated the implementation to work with the rails 5 attribute API we use in
0.7.