Introduce deprecation warnings for 0.6->0.7 differences
This means giving people a warning that `type` is meaningless on
`vault_attribute_proxy` now and they should move their type information to
the `vault_attribute` definition.  This also means reintroducing the
`Vault::AttributeProxy` module and having it do nothing but emit a
deprecation warning telling you to stop including it.