Actually drop support for rails 4.x
When we initially created version 0.7 of the gem to use the activerecord
attributes API, our goal was to create a unified version that worked with
rails 4.x and 5.x.  We ultimately failed in that because the attribute
API is private in rails 4.x and does not have the exact same behaviour as
it does in rails 5.x.  We decided that version 0.6 of the gem would support
rails 4.x and version 0.7 would support rails 5.x.  We didn't actually make
this choice explicit though, and so now we do.  This allows us to use rails
5.x features properly.