Detect serializer based on type for vault_attribute
For most cases, if we have specified a `type` option on `vault_attribute`
we will also need a serializer to deal with sending the value to vault as
a string, and converting it back again.  For most cases, we can also detect
the serializer based on the type and this means the author of the
`vault_attribute` definitions doesn't need to worry about it.