Use `ActiveRecord::Type.lookup` to find type classes
Our original strategy of looking for `ActiveRecord::Type::<the type>` works
for all the default rails types, but it doesn't allow us to make use of any
custom types that our database adapter might introduce.  In rails 5+
`ActiveRecord::Type.lookup` was introduced to let us find a type
implementation that has previously been registered.  This means we can use
the same set of types as our database columns, but also any custom types
that we might register ourselves (like a Money class or whatever).

This change is why we finally dropped support for rail 4.x as it does not
have this functionality so it's impossible to know what other custom types
are available at runtime.