Add a serializer for IPAddr objects
In postgresql there are inet and cidr types for columns that are handled by
the `ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Cidr` and
`ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Inet` type classes.
This means we get ruby `IPAddr` objects as the value for our attributes
when we read them out of the database with `ActiveRecord`.  We want to
support these column types out of the box so we've introduced a serializer
for `IPAddr` (based heavily on the implementation of the type class) and
set this serializer as the default for the `inet` and `cidr` types. We also
want to support `IPAddr` serialization for other column types that might
use `IPAddr` so we've also added it as the serializer for `ipaddr`, so that
other database adapters are not beholden to the names postgresql uses.