Add support for key derivation with a context
This adds support for passing a context to Vault Transit encrypt/decrypt
operations, allowing an application to specify a context per record or a
`belongs_to` relation (such as a user, team, or organization). The
context can be a string, or if given a proc or symbol, will generate the
context for each encrypt/decrypt request.

This changes the signature of `Vault::Rails.encrypt` and
`Vault::Rails.decrypt`. Each had an arity of 3-4, and they now have an
arity of 3 and 2 optional keyword arguments. The minor version number
should be bumped on release (since the major version is still 0).