Preserve empty JSON types
The existing behavior when specifying `serializer: :json` was to store
and return an empty Hash (`{}`) for null or empty values, including
empty arrays.

This changes the encode logic to store the JSON representation of
any given value. During decode, any valid JSON string can be decoded. We
special case `nil` or `""` (empty string) to return `nil` (instead of
`{}` previously).

This could be a breaking for applications that depend on the empty hash
== nil behavior.