Require a default with the built-in serializer
When upgrading from v0.4.0, the following attribute would have
previously returned an empty hash (`{}`) would now return `nil` for new
values.

```rb
vault_attribute :data,
  serialize: :json
```

In the changelog, we'll call out that this should have a default added
with `{}` to preserve legacy behavior, like so:

```rb
vault_attribute :data,
  serialize: :json,
  default: {}
```

This PR proposes to make the default a requirement for using built-in
serializers (of which only :json exists). This would make it more
apparent during local development and test runs that the default must be
added to existing serialized vault_attributes, via an error being raised
at model load time.

The downside of this change is that for attributes that do not require a
default, a default must still be passed:

```rb
vault_attribute :data,
  serialize: :json,
  default: nil
```

But perhaps requiring the user to be explicit is better than implicit in
this case.