Remove travis-ci dependency by moving completely to CircleCI
The "wwtd" rubygem along with Travis CI configuration file basically run the suite test against different versions of Ruby and Rails.
This causes problems when the ruby/rails combination are ruby versions lower than 2.5 and the rails version is 6 since this version of rails only supports ruby version 2.5.0 or higher.