Fix dirty nil assignments
Fixes a regression caused by
https://github.com/hashicorp/vault-rails/pull/99/commits/6c5b1a3c2a6e1f9e7e7f228dee9ce6e1717ce548

The usage of instance_variable_get means that if an attribute is
assigned to `nil`, the variable to signal that attributes are loaded is
`false` - meaning that any intentional setting of `nil` results in
attributes getting reloaded on the next read and wiping that `nil` value
back to whatever it was before. I've added a spec to describe this
behavior.