allow lazy decrypted vault attributes to be referenced after a destroy
After destroying a record, ActiveRecord marks the in-memory object as
frozen and unable to be altered. This is fine, unless lazy decrypt is
enabled and you need to access an encrypted attribute after the destroy
happens - for instance, as part of a dependent: :destroy chain.
In that case, vault-rails will attempt to set the plaintext attribute
_after_ the destroy, raising a "Can't modify frozen hash" error.

This commit ensures that encrypted attributes are always decrypted
prior to a destroy.